requirements.txt
# Core Dependencies
Flask==3.0.0
python-dotenv==1.0.0

# AI & LLM
groq>=0.11.0

# Google APIs
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Database (SQLite is built-in to Python)

# Optional but recommended
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0